Refactor file util
