+++ lib/streaming_response.py
@@ -3,10 +3,12 @@ from pygments import highlight
 from pygments.lexers import get_lexer_by_name
 from pygments.formatters import TerminalFormatter
 from pygments.util import ClassNotFound
-from lib.file_util import (
+from lib.file_parser import (
+    parse_files,
+)
+from lib.file_parser import (
     extract_filename_start,
     extract_filename_end,
-    parse_files,
 )
 from lib.shell_util import (
     LIGHT_PINK,
